requirements.txt
mysql-connector-python==9.6.0
confluent-kafka==2.13.0
neo4j==6.1.0
python-dotenv==1.2.2
tzdata==2025.3
pydantic==2.12.5
